+++ src/render/backends/sw/mod.rs
@@ -405,7 +405,7 @@ impl<'a> Canvas for SwRenderer<'a> {
     fn draw_posed_glyph_run(
         &mut self,
         pos: &Point,
-        glyphs: crate::render::command::PosedGlyphs<'_>,
+        glyphs: crate::render::PosedGlyphs<'_>,
         font: &crate::render::font::Font,
         clip: &Rect,
         color: &Color,